backend/requirements.txt

fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
pymupdf==1.23.21
python-docx==1.1.0
docx2pdf==0.1.8
pillow==10.2.0
pytesseract==0.3.10
requests==2.31.0
python-dotenv==1.0.1
